Implemented args() and cflow(). args() (and binding in general) doesn't handle alpha-renaming for pointcut references yet.

+++ edu.ubc.mirrors.jdi/src/edu/ubc/mirrors/jdi/JDIVirtualMachineMirror.java
@@ -298,7 +298,7 @@ public class JDIVirtualMachineMirror implements VirtualMachineMirror {
 	} else if (value instanceof FloatValue) {
 	    return ((FloatValue)value).floatValue();
 	} else if (value instanceof DoubleValue) {
-	    return ((FloatValue)value).doubleValue();
+	    return ((DoubleValue)value).doubleValue();
 	} else {
 	    return makeMirror((ObjectReference)value);
 	}
